mshv-ioctls: don't set TB flush hypercall feature for arm64 guests
This synthetic processor feature is not supported for arm64 in older
hypervisor builds resulting in partition creation faiures. It is
important to support this older hypervisor build (AH24) because it
is currently live in the Azure fleet.

This older build doesn't even support L1VH parent self querying
synthetic processor features to detect which features are supported
by the host.

For now, skip this feature entirely for arm64 guests. This is okay
because Linux guests don't even use these hypercalls.

Signed-off-by: Anirudh Rayabharam <anrayabh@microsoft.com>